requirements.txt
# ===================================
# VocaNote - Dépendances Python
# ===================================
# Transcription, Diarisation, Résumé IA, Licence

# === INTERFACE ===
PyQt6>=6.6.0

# === TRANSCRIPTION (Whisper) ===
openai-whisper

# === MACHINE LEARNING (PyTorch) ===
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0

# === DIARISATION (Pyannote) ===
pyannote.audio>=3.1.1
pyannote.core>=5.0.0
pyannote.metrics>=3.2.1
pyannote.pipeline>=3.0.1
pyannote.database>=5.0.1

# === RÉSUMÉ (Transformers/BART) ===
transformers>=4.30.0
sentencepiece>=0.1.99
tokenizers>=0.22.0,<0.24.0
safetensors>=0.4.0
accelerate>=0.20.0

# === HUGGINGFACE ===
huggingface-hub==0.23.0

# === AUDIO ===
av>=10.0.0
ffmpeg-python
imageio-ffmpeg
soundfile>=0.12.0

# === TRAITEMENT ===
scipy>=1.10.0
regex>=2023.0.0
tqdm>=4.65.0

# === LICENCE / SÉCURITÉ ===
cryptography>=41.0.0

# === BUILD (optionnel) ===
# pyinstaller>=6.0.0
